Move ClockUtil to correct package
