+++ amplify/src/main/java/com/github/stkent/amplify/utils/ClockUtil.java
@@ -14,7 +14,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package com.github.stkent.amplify.tracking;
+package com.github.stkent.amplify.utils;
 
 import android.support.annotation.Nullable;
